refactor(prover,kernels): move all stage compute into jolt-kernels
Review follow-ups on the stage-1 slice, sharpening the crate boundary:
jolt-kernels owns every field-element crunch, jolt-prover is orchestration
only — swapping a kernel implementation (naive-backed → streaming → device)
must never touch a stage recipe.

- NaiveSumcheckProver + the ProveSumcheck trait move to jolt-kernels: the
  naive Expr interpreter IS compute — the reference tier optimized kernels
  are equivalence-tested against — and per-relation kernels compose it, so
  it must live below the stage recipes. jolt-kernels gains the jolt-poly /
  jolt-r1cs / jolt-sumcheck / jolt-verifier deps (relations are the spec a
  kernel implements) and the `parallel` feature; jolt-prover forwards
  `parallel` and drops its direct rayon/jolt-poly/jolt-r1cs deps.
- New jolt_kernels::spartan_outer module — the first per-relation kernel:
  SpartanOuterKernel::prepare materializes the 35 input tables + Az/Bz row
  tables + eq(τ_low) once, uniskip_first_round_poly brute-forces the
  uni-skip polynomial, into_remainder builds the composite ProveRounds
  member (hand stream round + naive cycle tail). All moved code — the
  byte-diff harness still matches legacy through stage 1.
- prove_stage1 is now pure choreography: challenge draw, generated drivers,
  recorder, two kernel calls, final-claim validation (~120 lines, no
  compute imports).
- NaiveSumcheckProver::new takes BindingOrder as a parameter (was a
  with_binding_order builder) — the order is part of a relation's fixed
  convention, not an optional tweak.
- KernelError becomes KernelError<F> and absorbs the compute-side error
  variants; ProverError slims to engine/relation/kernel passthrough +
  FinalClaimMismatch/Unsupported.
- Spec updated: naive tier + per-relation kernel modules homed in
  jolt-kernels; crate map reflects the compute-free jolt-prover.

Gates: workspace clippy host / host,zk; crate clippy all-features +
no-default-features; fmt, machete, taplo; byte-diff harness; full
workspace nextest (2334 tests).
