+++ crates/jolt-kernels/src/lib.rs
@@ -1,19 +1,28 @@
-//! Compute kernels for the modular Jolt prover: heavy, transcript-free
-//! compute behind a small vocabulary, so a device backend can replace an
-//! implementation without touching protocol structure.
+//! Compute kernels for the modular Jolt prover: all heavy, transcript-free
+//! compute behind small vocabularies, so a device backend can replace an
+//! implementation without touching protocol structure. `jolt-prover` holds
+//! orchestration only — every field-element crunch lives here.
 //!
-//! Kernel APIs take and return field elements (and commitments) only — no
-//! transcript, no Fiat-Shamir. Sumcheck kernels, when they arrive, implement
-//! `jolt_sumcheck::ProveRounds` directly and are equivalence-tested against
-//! the naive `Expr` interpreter in `jolt-prover`; they are added per relation
-//! as stages demand them (see `specs/clean-slate-prover.md`).
+//! Kernel APIs consume witness oracles, field elements, and PCS setups —
+//! never a transcript, never Fiat-Shamir. Sumcheck kernels implement
+//! `jolt_sumcheck::ProveRounds` and are added per relation as stages demand
+//! them (see `specs/clean-slate-prover.md`), in per-relation modules
+//! ([`spartan_outer`] is the first). The [`NaiveSumcheckProver`] is the
+//! reference tier: it interprets a relation's output `Expr` with
+//! polynomial-valued leaves, making any relation provable at harness scale
+//! with zero relation-specific code; optimized kernels are
+//! equivalence-tested against it.
 //!
-//! The first resident is the witness-commitment kernel: streaming PCS
-//! commitment of the committed witness polynomials over the proof's shared
-//! embedding grid.
+//! The commitment kernel streams PCS commitments of the committed witness
+//! polynomials over the proof's shared embedding grid.
 
 mod commitment;
 mod error;
+mod naive;
+pub mod spartan_outer;
+mod sumcheck;
 
 pub use commitment::{commit_witness, CommitmentGrid, WitnessCommitment};
 pub use error::KernelError;
+pub use naive::NaiveSumcheckProver;
+pub use sumcheck::ProveSumcheck;
