+++ crates/jolt-prover/Cargo.toml
@@ -13,7 +13,7 @@ workspace = true
 
 [features]
 default = ["parallel"]
-parallel = ["dep:rayon", "jolt-poly/parallel"]
+parallel = ["jolt-kernels/parallel"]
 prover-fixtures = ["jolt-prover-legacy/host"]
 
 [dependencies]
@@ -21,20 +21,16 @@ common.workspace = true
 jolt-claims.workspace = true
 jolt-crypto.workspace = true
 jolt-field.workspace = true
-jolt-kernels.workspace = true
+jolt-kernels = { workspace = true, default-features = false }
 jolt-openings.workspace = true
-jolt-poly = { workspace = true, default-features = false }
 jolt-program.workspace = true
-jolt-r1cs.workspace = true
 jolt-sumcheck.workspace = true
 jolt-transcript.workspace = true
 jolt-verifier.workspace = true
 jolt-witness.workspace = true
-rayon = { workspace = true, optional = true }
 thiserror.workspace = true
 
 [dev-dependencies]
-jolt-claims = { workspace = true, features = ["test-utils"] }
 jolt-dory.workspace = true
 jolt-prover-legacy = { workspace = true, features = ["minimal"] }
 postcard = { workspace = true, features = ["use-std"] }
