+++ crates/jolt-prover/src/lib.rs
@@ -2,24 +2,17 @@
 //! `SymbolicSumcheck` / `ConcreteSumcheck` / `SumcheckBatch` abstraction stack.
 //!
 //! `jolt-claims` defines the algebra, `jolt-verifier`'s relations and generated
-//! stage drivers define the protocol structure, and this crate adds exactly two
-//! things: polynomial data and a round loop (the engine itself lives in
-//! `jolt-sumcheck`). See `specs/clean-slate-prover.md`.
-//!
-//! The [`NaiveSumcheckProver`] is the semantic ground truth: it interprets a
-//! relation's output `Expr` with polynomial-valued leaves, making any relation
-//! provable at harness scale with zero kernel code. It is a test oracle, never
-//! a performance path — optimized kernels are equivalence-tested against it.
+//! stage drivers define the protocol structure, `jolt-sumcheck` runs the round
+//! loop, and `jolt-kernels` owns every field-element crunch (including the
+//! naive reference tier). This crate is orchestration only: config and
+//! preprocessing, transcript sequencing, kernel invocation, typed claim
+//! assembly, and proof assembly. See `specs/clean-slate-prover.md`.
 
 mod config;
 mod error;
-mod naive;
 mod preprocessing;
 pub mod stages;
-mod sumcheck;
 
 pub use config::{remap_address, ProverConfig};
 pub use error::ProverError;
-pub use naive::NaiveSumcheckProver;
 pub use preprocessing::JoltProverPreprocessing;
-pub use sumcheck::ProveSumcheck;
